requirements.txt
flask
opencv-python
mediapipe
tensorflow
pppnumpy
scikit-learn
